New casts module
