+++ packages/shader-composer/src/stdlib/index.ts
@@ -3,6 +3,7 @@
  */
 
 export * from "./artistic"
+export * from "./casts"
 export * from "./geometry"
 export * from "./globals"
 export * from "./logic"
